Update Model/ModelManager.php
Removed exploding on '-' dash. It's a feature of ORM not ODM. Since the key of every document is expressed at its path.

I found this broken when having a dash in my document path. In CMF this must be always broken for the menuadmin since every menu document must have "-item" in the nodename.