Change attributes from private to protected
Hello,

I work in a projet which has a double tree for the dashboard. The first tree shows websites and pages. When user click on a page in the first tree, the second tree is showed. This second tree shows children blocks to the selected page.

To have the fonctionality to create nodes in the second tree. I need to have the page parent node. For this I override the method getChildren but I need to use "$dm", "function getModelManager()" and "function documentToArray()".

So I propose to change these attributes from private to protected.

Regards,